Extract useDocumentSerialization and useFileHandling composables from EditorApp.vue
Move serialization/cache logic (serializeDocument, deserializeToDocument,
scheduleCacheSave, handleSlotSwitch) and file I/O logic (readFile with
MAP/JSON parsing) into dedicated composables, reducing EditorApp.vue
from 1,933 to 1,586 lines.
